+++ lib/active_date_range/active_model_type.rb
@@ -2,8 +2,17 @@
 
 module ActiveDateRange
   class DateRangeType < ActiveModel::Type::String
+    # Pass <tt>safe: true</tt> for attributes fed by input you don't control, like a query
+    # parameter. The attribute then casts to nil instead of raising:
+    #
+    #   attribute :period, :date_range, safe: true
+    def initialize(safe: false)
+      @safe = safe
+      super()
+    end
+
     def cast(value)
-      ActiveDateRange::DateRange.parse(value)
+      @safe ? ActiveDateRange::DateRange.safe_parse(value) : ActiveDateRange::DateRange.parse(value)
     end
   end
 end